Refactor:index.html과 index.js, style.css에 몰려있는 코드를 분할
