+++ components/header.html
@@ -1,0 +1,24 @@
+<header class="header">
+    <!-- 왼쪽 로고 -->
+    <a href="index.html" class="logo">Wise Wallet</a>
+
+    <!-- ✅ 중앙 블럭 -->
+    <div class="center-block">
+        <div class="month-selector">
+        <div class="year" id="year">2023</div>
+        <div class="month-control">
+            <button id="prev-month" class="nav-btn">&lt;</button>
+            <div class="month" id="month">8</div>
+            <button id="next-month" class="nav-btn">&gt;</button>
+        </div>
+        <div class="month-label" id="month-label">August</div>
+        </div>
+    </div>
+
+    <!-- 오른쪽 보기 버튼 -->
+    <div class="view-options">
+        <button class="icon-btn active">📄</button>
+        <button class="icon-btn">📅</button>
+        <button class="icon-btn">📊</button>
+    </div>
+</header>
