+++ components/input-form.html
@@ -1,0 +1,62 @@
+<section class="input-form">
+            <div class="inter-input-block">
+                <div>일자</div>
+                <input type="date" id="date" />
+            </div>
+
+            <div class="vertical-line"></div>
+
+            <div class="inter-input-block">
+                <div>금액</div>
+                <div class ="amount-block">
+                    <div class="toggle-sign" id="toggle-sign">+</div>
+                    <input type="text" id="amount" placeholder="0" />
+                    <div>원</div>
+                </div>
+            </div>
+
+            <div class="vertical-line"></div>
+
+            <div class="inter-input-block">
+                <div class="desc-header">
+                    <div>내용</div>
+                    <div class="char-count" id="char-count">0 / 32</div>
+                </div>
+            <input type="text" id="desc" class="content" placeholder="입력하세요" maxlength="32" />
+            </div>
+
+            <div class="vertical-line"></div>
+
+            <div class="inter-input-block" id="method-wrapper">
+                <div class="method">결제수단</div>
+                <div class="dropdown-display" id="dropdown-display">선택하세요</div>
+                <div class="dropdown-panel hidden" id="dropdown-panel">
+                    <div class="dropdown-add" id="dropdown-add">추가하기</div>
+                </div>
+                </div>
+
+
+            <!-- 모달 -->
+            <div id="modal" class="modal hidden">
+            <div class="modal-content">
+                <h3>결제수단 추가</h3>
+                <input type="text" id="new-method" placeholder="예: 우리카드" />
+                <button id="confirm-add">추가하기</button>
+                <button id="cancel-add">취소</button>
+            </div>
+            </div>
+
+            <div class="vertical-line"></div>
+
+            <div class="inter-input-block" id="category-wrapper">
+                <div>분류</div>
+                <div class="dropdown-display" id="category-display">선택하세요</div>
+                <div class="dropdown-panel hidden" id="category-panel">
+                    <!-- 수입/지출 항목은 JS로 렌더링 -->
+                </div>
+            </div>
+
+            <button id="add-btn">
+                <img src="assets/botton.png" alt="확인" width="20" height="20" />
+            </button>
+            </section>
