+++ css/style.css
@@ -6,141 +6,8 @@ body {
   font-family: 'Segoe UI', sans-serif;
 }
 
-/* 헤더 전체 */
-.header {
-  height: 200px;
-  position: relative;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background: #73A4D0;
-  padding: 16px 400px;
-  border-bottom: 1px solid #ddd;
-}
 
 
-
-.month-selector {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 4px;
-}
-
-.month-control {
-    display: flex;
-    align-items: center;
-    justify-content : center;
-    gap:16px;
-    font-size:1.8rem;
-    font-weight:bold;
-}
-
-/* 왼쪽 로고 */
-.logo {
-  font-size: 1.5rem;
-  font-weight: bold;
-  color: #333;
-  text-decoration: none;
-}
-
-/* 가운데 월 선택 */
-.month-nav {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  font-size: 1.2rem;
-  color: #444;
-}
-
-.nav-btn {
-  background: none;
-  border: none;
-  font-size: 1.2rem;
-  cursor: pointer;
-  padding: 4px 8px;
-}
-
-/* 오른쪽 뷰 옵션 */
-.view-options {
-  display: flex;
-  gap: 8px;
-}
-
-.icon-btn {
-  background: none;
-  border: none;
-  font-size: 1.2rem;
-  cursor: pointer;
-  padding: 6px;
-}
-
-.icon-btn.active {
-  background: #fff;
-  color: #000;
-  font-weight: bold;
-  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.15);
-}
-
-
-.input-form {
-  position: relative; /* 또는 absolute도 가능 */
-  top: -30px;          /* 헤더를 위로 겹쳐서 올라가게 */
-  margin: 0 auto;
-  padding: 16px;
-  width: fit-content;
-  background: #fff;
-  border: 1px solid #000;
-  display: flex;
-  gap: 8px;
-  align-items: center;
-  z-index: 10; /* 헤더 위에 떠야 한다면 1 이상 */
-}
-
-.vertical-line {
-  width: 1px;
-  height: 48px;
-  background-color: #ccc;
-  align-self: center;
-}
-
-.amount-block{
-    display: flex;
-    justify-content: space-between;
-    gap: 8px;
-}
-
-#amount {
-  border: none;
-  background: transparent;
-  outline: none;
-  text-align: right;
-  width: 100px;
-  color: #000;
-}
-
-#amount::placeholder {
-  color: #aaa;
-}
-
-.toggle-sign {
-  width: 12px;
-  height: 22px;
-  border-radius: 50%;
-  color: #000;
-  font-weight: bold;
-  font-size: 20px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  user-select: none;
-}
-
-
-.inter-input-block{
-    position: relative;
-}
 
 .entry-date-section{
   margin : 20px 400px;
@@ -222,101 +89,3 @@ body {
 }
 
 
-.content{
-    border: none;
-}
-
-
-.desc-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-}
-
-.char-count {
-  font-size: 12px;
-  color: #888;
-}
-
-#date{
-    border: none;
-}
-
-/* 선택 영역 */
-.dropdown-display {
-  border: none;
-  cursor: pointer;
-  color: #ccc;
-  font-size: 12px;
-  width: 100px;
-  margin: 4px;
-}
-
-/* 드롭다운 목록 */
-.dropdown-panel {
-  position: absolute;
-  top: 58px;
-  left: -10px;
-  width: 120%;
-  background: white;
-  border: 1px solid #ccc;
-  margin-top: 4px;
-  z-index: 10;
-}
-
-.dropdown-option,
-.dropdown-add {
-  padding: 10px;
-  cursor: pointer;
-}
-
-.dropdown-option:hover,
-.dropdown-add:hover {
-  background-color: #f0f0f0;
-}
-
-.hidden {
-  display: none;
-}
-
-.dropdown-option {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.delete-btn {
-  background: none;
-  border: none;
-  cursor: pointer;
-  font-size: 14px;
-  color: red;
-  padding: 0 4px;
-}
-
-/* 모달 재사용 */
-.modal {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  background: rgba(0, 0, 0, 0.4);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 999;
-}
-
-.modal.hidden {
-  display: none;
-}
-
-.modal-content {
-  background: white;
-  padding: 20px;
-  border-radius: 8px;
-  width: 280px;
-  text-align: center;
-}
