+++ css/total-amount.css
@@ -1,0 +1,8 @@
+.total-amount {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin: 20px 0;
+    font-size: 1.5rem;
+    color: #333;
+}
