+++ index.html
@@ -4,96 +4,97 @@
         <meta charset="UTF-8">
         <title>Wise Wallet</title>
         <link rel="stylesheet" href="css/style.css" />
+        <link rel="stylesheet" href="css/header.css" />
+        <link rel="stylesheet" href="css/input-form.css" />
     </head>
     <body>
-        <header class="header">
-            <!-- 왼쪽 로고 -->
-            <a href="index.html" class="logo">Wise Wallet</a>
-
-            <!-- ✅ 중앙 블럭 -->
-            <div class="center-block">
-                <div class="month-selector">
-                <div class="year" id="year">2023</div>
-                <div class="month-control">
-                    <button id="prev-month" class="nav-btn">&lt;</button>
-                    <div class="month" id="month">8</div>
-                    <button id="next-month" class="nav-btn">&gt;</button>
+        <!-- 헤더부분입니다 -->
+        <div id="header-container"></div>
+        <!-- 입력 폼 부분입니다 -->
+        <div id="input-form-container"></div>
+        <div class="total-amount">
+            <div class="total-amount-text">총 금액</div>
+            <div class="total-amount-value" id="total-amount-value">0 원</div>
+        </div>
+        <section class="entry-list" id="entry-list">
+            <!-- 더미 데이터 시작 -->
+            <div class="entry-date-section" data-date="2025-07-10">
+                <div class="entry-sort">
+                    <div>7월 10일 목요일</div>
+                    <div>235,000</div>
                 </div>
-                <div class="month-label" id="month-label">August</div>
-                </div>
-            </div>
-
-            <!-- 오른쪽 보기 버튼 -->
-            <div class="view-options">
-                <button class="icon-btn active">📄</button>
-                <button class="icon-btn">📅</button>
-                <button class="icon-btn">📊</button>
-            </div>
-        </header>
-
-        <section class="input-form">
-            <div class="inter-input-block">
-                <div>일자</div>
-                <input type="date" id="date" />
-            </div>
-
-            <div class="vertical-line"></div>
-
-            <div class="inter-input-block">
-                <div>금액</div>
-                <div class ="amount-block">
-                    <div class="toggle-sign" id="toggle-sign">+</div>
-                    <input type="text" id="amount" placeholder="0" />
-                    <div>원</div>
+                <div class="entry-items">
+                    <div class="entry-row">
+                        <div class="entry-category">식비</div>
+                        <div class="entry-desc">점심 식사</div>
+                        <div class="entry-method">신용카드</div>
+                        <div class="entry-amount expense-amount">-15,000원</div>
+                    </div>
+                    <div class="entry-row">
+                        <div class="entry-category">쇼핑/뷰티</div>
+                        <div class="entry-desc">여름 옷 구매</div>
+                        <div class="entry-method">체크카드</div>
+                        <div class="entry-amount expense-amount">-120,000원</div>
+                    </div>
+                    <div class="entry-row">
+                        <div class="entry-category">월급</div>
+                        <div class="entry-desc">7월 월급</div>
+                        <div class="entry-method">계좌이체</div>
+                        <div class="entry-amount income-amount">+3,000,000원</div>
+                    </div>
                 </div>
             </div>
-
-            <div class="vertical-line"></div>
-
-            <div class="inter-input-block">
-                <div class="desc-header">
-                    <div>내용</div>
-                    <div class="char-count" id="char-count">0 / 32</div>
+            
+            <div class="entry-date-section" data-date="2025-07-09">
+                <div class="entry-sort">
+                    <div>7월 9일 수요일</div>
+                    <div>52,000</div>
                 </div>
-            <input type="text" id="desc" class="content" placeholder="입력하세요" maxlength="32" />
-            </div>
-
-            <div class="vertical-line"></div>
-
-            <div class="inter-input-block" id="method-wrapper">
-                <div class="method">결제수단</div>
-                <div class="dropdown-display" id="dropdown-display">선택하세요</div>
-                <div class="dropdown-panel hidden" id="dropdown-panel">
-                    <div class="dropdown-add" id="dropdown-add">추가하기</div>
-                </div>
-                </div>
-
-
-            <!-- 모달 -->
-            <div id="modal" class="modal hidden">
-            <div class="modal-content">
-                <h3>결제수단 추가</h3>
-                <input type="text" id="new-method" placeholder="예: 우리카드" />
-                <button id="confirm-add">추가하기</button>
-                <button id="cancel-add">취소</button>
-            </div>
-            </div>
-
-            <div class="vertical-line"></div>
-
-            <div class="inter-input-block" id="category-wrapper">
-                <div>분류</div>
-                <div class="dropdown-display" id="category-display">선택하세요</div>
-                <div class="dropdown-panel hidden" id="category-panel">
-                    <!-- 수입/지출 항목은 JS로 렌더링 -->
+                <div class="entry-items">
+                    <div class="entry-row">
+                        <div class="entry-category">교통</div>
+                        <div class="entry-desc">택시비</div>
+                        <div class="entry-method">카카오페이</div>
+                        <div class="entry-amount expense-amount">-12,000원</div>
+                    </div>
+                    <div class="entry-row">
+                        <div class="entry-category">문화/여가</div>
+                        <div class="entry-desc">영화 관람</div>
+                        <div class="entry-method">신용카드</div>
+                        <div class="entry-amount expense-amount">-15,000원</div>
+                    </div>
+                    <div class="entry-row">
+                        <div class="entry-category">식비</div>
+                        <div class="entry-desc">저녁 식사</div>
+                        <div class="entry-method">현금</div>
+                        <div class="entry-amount expense-amount">-25,000원</div>
+                    </div>
                 </div>
             </div>
-
-            <button id="add-btn">
-                <img src="assets/botton.png" alt="확인" width="20" height="20" />
-            </button>
-            </section>
-        <section class="entry-list" id="entry-list"></section>
-        <script src="js/middleCalendar.js"></script>
+            
+            <div class="entry-date-section" data-date="2025-07-08">
+                <div class="entry-sort">
+                    <div>7월 8일 화요일</div>
+                    <div>75,000</div>
+                </div>
+                <div class="entry-items">
+                    <div class="entry-row">
+                        <div class="entry-category">용돈</div>
+                        <div class="entry-desc">생일 축하 용돈</div>
+                        <div class="entry-method">계좌이체</div>
+                        <div class="entry-amount income-amount">+100,000원</div>
+                    </div>
+                    <div class="entry-row">
+                        <div class="entry-category">의료/건강</div>
+                        <div class="entry-desc">약국</div>
+                        <div class="entry-method">체크카드</div>
+                        <div class="entry-amount expense-amount">-25,000원</div>
+                    </div>
+                </div>
+            </div>
+            <!-- 더미 데이터 끝 -->
+        </section>
+        <script type="module"src="js/loadComponents.js"></script>
+        <!-- <script type="module" src="js/middleCalendar.js"></script> -->
     </body>
 </html>
